removed warnings from compiled code except auto layout warnings, seems to be a little bug in Xcode

+++ UniversalExpenseTracker/MasterViewController.m
@@ -1177,15 +1177,15 @@ typedef NS_ENUM(NSInteger, UETFilterType) {
 #pragma mark - XYPieChart Delegate
 - (void)pieChart:(XYPieChart *)pieChart willSelectSliceAtIndex:(NSUInteger)index
 {
-    NSLog(@"will select slice at index %d",index);
+    NSLog(@"will select slice at index %lu",(unsigned long)index);
 }
 - (void)pieChart:(XYPieChart *)pieChart willDeselectSliceAtIndex:(NSUInteger)index
 {
-    NSLog(@"will deselect slice at index %d",index);
+    NSLog(@"will deselect slice at index %lu",(unsigned long)index);
 }
 - (void)pieChart:(XYPieChart *)pieChart didDeselectSliceAtIndex:(NSUInteger)index
 {
-    NSLog(@"did deselect slice at index %d",index);
+    NSLog(@"did deselect slice at index %lu",(unsigned long)index);
 }
 //- (void)pieChart:(XYPieChart *)pieChart didSelectSliceAtIndex:(NSUInteger)index
 //{
